Logica parecida del update detail en destroy detail

+++ app/Http/Controllers/OrderPurchaseController.php
@@ -969,13 +969,16 @@ class OrderPurchaseController extends Controller
 
             $material_order = MaterialOrder::where('material_id', $idMaterial)
                 ->where('order_purchase_detail_id', $detail->id)->first();
-
-            if ( $material_order->quantity_entered > 0 ) {
-                return response()->json(['message' => 'No se puede eliminar el detalle porque ya hay un ingreso.'], 422);
-            } else {
-                $material_order->delete();
-                $detail->delete();
+            if ( isset($material_order) )
+            {
+                if ( $material_order->quantity_entered > 0 ) {
+                    return response()->json(['message' => 'No se puede eliminar el detalle porque ya hay un ingreso.'], 422);
+                } else {
+                    $material_order->delete();
+                    $detail->delete();
+                }
             }
+
             DB::commit();
         } catch ( \Throwable $e ) {
             DB::rollBack();
